Added quest function RunTShockCommand(QPlaye player, string commandText)
Directly uses the TShockAPI.Commands.HandleCommand function. player is
the current quest player, commandText is a full command message as if
the player were issuing it. Ex: "/kick Ijwu"

This isn't what I first planned, but it's a good temporary
implementation. It DOES take into account permissions. So if a quest
tries to execute a command a player has no access to, they will NOT be
able to continue. (They will just get the typical TShock "You don't have
permission..." message.)